% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{pm_explorer}
\alias{pm_explorer}
\title{Launch the Pmetrics Explorer app}
\usage{
pm_explorer(launch.browser = TRUE, data_env = .GlobalEnv, ...)
}
\arguments{
\item{launch.browser}{Logical, passed to [shiny::runApp()].}

\item{data_env}{An environment to search for PM_model and PM_data objects.
Defaults to [.GlobalEnv].}

\item{...}{Additional arguments passed to golem options.}
}
\description{
Convenience wrapper around [run_app()] to rapidly simulate from a
Pmetrics model and explore parameter ranges interactively.
}
