% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(launch.browser = TRUE, data_env = .GlobalEnv, ...)
}
\arguments{
\item{launch.browser}{Logical, passed to [shiny::runApp()].}

\item{data_env}{An environment to search for PM_model and PM_data objects.
Defaults to [.GlobalEnv].}

\item{...}{Additional arguments passed to golem options.}
}
\description{
Run the Shiny Application
}
